requirements.txt
folium==0.14.0
haversine==2.7.0
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
plotly==5.9.0
streamlit==1.17.0
streamlit_folium==0.11.0
